Moved patches and fixed Scp173FirstKillPatch
